refactor: move MultilinePasteWidget
